Get last location

+++ app/src/main/java/com/yusmle/restaurants/features/restaurantslist/business/RestaurantRepository.kt
@@ -6,8 +6,8 @@ interface RestaurantRepository {
 
     suspend fun fetchRemoteRestaurantsList(
         pagingMetaData: String?,
-        longitude: Float,
-        latitude: Float
+        longitude: Double,
+        latitude: Double
     ): RestaurantsSearchBundle
 
     suspend fun deletePersistedRestaurantsList()
